backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
supabase==2.9.1
PyJWT==2.9.0
cryptography==43.0.1
numpy==2.1.1
grpcio==1.68.0
